Rename some folders and update examples.
