+++ deep_speech_2/examples/tiny/run_infer.sh
@@ -4,7 +4,7 @@ pushd ../..
 
 CUDA_VISIBLE_DEVICES=0 \
 python -u infer.py \
---num_samples=10 \
+--num_samples=4 \
 --trainer_count=1 \
 --beam_size=500 \
 --num_proc_bsearch=12 \
@@ -17,11 +17,11 @@ python -u infer.py \
 --use_gru=False \
 --use_gpu=True \
 --share_rnn_weights=True \
---infer_manifest='data/librispeech/manifest.dev-clean' \
---mean_std_path='data/librispeech/mean_std.npz' \
---vocab_path='data/librispeech/eng_vocab.txt' \
---model_path='checkpoints/params.latest.tar.gz' \
---lang_model_path='lm/data/common_crawl_00.prune01111.trie.klm' \
+--infer_manifest='data/tiny/manifest.train' \
+--mean_std_path='data/tiny/mean_std.npz' \
+--vocab_path='data/tiny/vocab.txt' \
+--model_path='checkpoints/params.pass-14.tar.gz' \
+--lang_model_path='models/lm/common_crawl_00.prune01111.trie.klm' \
 --decoding_method='ctc_beam_search' \
 --error_rate_type='wer' \
 --specgram_type='linear'
